+++ workspaces/adr/plugins/adr-backend/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @backstage-community/plugin-adr-backend
 
+## 0.5.0
+
+### Minor Changes
+
+- d5e54af: Backstage version bump to v1.34.1
+
+### Patch Changes
+
+- Updated dependencies [d5e54af]
+  - @backstage-community/plugin-adr-common@0.3.0
+  - @backstage-community/search-backend-module-adr@0.3.0
+
 ## 0.4.27
 
 ### Patch Changes